Restructure http crate (#276)
The plan consists of restructuring HTTP crates following best practices to simplify the understanding of each component.
