+++ huginn-net-http/src/http2/process.rs
@@ -1,8 +1,11 @@
+use super::parser as http2_parser;
 use crate::error::HuginnNetHttpError;
+use crate::http;
+use crate::http::common as http_common;
+use crate::http::common::HttpProcessor;
+use crate::http::languages as http_languages;
+use crate::http::observable::{ObservableHttpRequest, ObservableHttpResponse};
 use crate::http::Header;
-use crate::http_common::HttpProcessor;
-use crate::observable::{ObservableHttpRequest, ObservableHttpResponse};
-use crate::{http, http2_parser, http_common, http_languages};
 use tracing::debug;
 
 /// HTTP/2 Protocol Processor
