+++ huginn-net-http/tests/akamai_paper_golden.rs
